Add Babel as a transpiling option for JS bundles (Fixes #10467)
