+++ .eslintrc.js
@@ -4,7 +4,6 @@ module.exports = {
         'commonjs': true
     },
     extends: [
-        '@mozilla-protocol/eslint-config',
         'plugin:json/recommended'
     ],
     /**
@@ -12,18 +11,27 @@ module.exports = {
      * */
     overrides: [
         {
+            // JS files transpiled by Babel
             files: [
-                'media/js/firefox/welcome/**/*.js',
-                'media/js/firefox/whatsnew/**/*.js'
+                'media/js/**/*.es6.js',
             ],
             env: {
-                'es6': true
-            },
-            parserOptions: {
-                ecmaVersion: 8
+                'es2017': true
             }
         },
         {
+            // JS files served only to Firefox browsers.
+            files: [
+                'media/js/firefox/welcome/**/*.js',
+                'media/js/firefox/whatsnew/**/*.js',
+                'media/js/firefox/firstrun/**/*.js',
+            ],
+            env: {
+                'es2017': true
+            }
+        },
+        {
+            // JS build files for local dev.
             files: [
                 'webpack.config.js',
                 'webpack.static.config.js',
@@ -31,10 +39,7 @@ module.exports = {
             ],
             env: {
                 'node': true,
-                'es6': true
-            },
-            parserOptions: {
-                ecmaVersion: 8
+                'es2017': true
             },
             rules: {
                 'strict': ['error', 'global'],
@@ -43,6 +48,7 @@ module.exports = {
     ],
     globals: {
         'Mozilla': 'writable',
+        'Mzp': 'writable',
         'site': 'writable'
     }
 };
