+++ media/static-bundles.json
@@ -1555,7 +1555,7 @@
       "files": [
         "protocol/js/protocol-modal.js",
         "js/base/mozilla-lazy-load.js",
-        "js/foundation/annual-report-2019.js"
+        "js/foundation/annual-report-2019.es6.js"
       ],
       "name": "annual_report_2019"
     },
