+++ package.json
@@ -4,9 +4,11 @@
   "description": "Making mozilla.org awesome, one pebble at a time",
   "private": true,
   "dependencies": {
+    "@babel/core": "^7.15.5",
+    "@babel/preset-env": "^7.15.6",
     "@mozilla-protocol/core": "14.0.3",
-    "@mozilla-protocol/eslint-config": "^1.1.0",
     "@sentry/browser": "^6.12.0",
+    "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "9.0.1",
     "css-loader": "6.2.0",
