+++ webpack.config.js
@@ -51,6 +51,28 @@ const jsConfig = {
         path: path.resolve(__dirname, 'assets/'),
         publicPath: '/media/',
     },
+    module: {
+        rules: [
+            {
+                test: /\.es6\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: [
+                            [
+                                '@babel/preset-env', {
+                                    targets: {
+                                        ie: '10'
+                                    }
+                                }
+                            ]
+                        ]
+                    }
+                }
+            }
+        ]
+    },
     watchOptions: {
         aggregateTimeout: 600,
         ignored: './node_modules/'
